requirements.txt
numpy
sounddevice
wavio
pyttsx3
google-generativeai
faster-whisper
openai-whisper
torch
requests
python-dotenv
sarvamai
python-docx
rich